Add files via upload
Changed a bit of the styling. Put the form component on the item being edited instead of just returning the form component and not showing the rest of the list. Changed the placeholder text for the form when it's being rendered for editing an item, it now says 'Edit text "--text of item being edited here--"'. Also changed the text for the button on the form component to say 'submit edit' when being rendered for editing.